Fix segment base resolution to store translated host pointers

+++ Android/app/src/main/cpp/emulator/gfx_interpreter.h
@@ -65,7 +65,7 @@ typedef struct {
     int matrixMode;  // 0=modelview, 1=projection
 
     // RSP segment base addresses (F3DEX_GBI)
-    uint32_t segmentBase[16];
+    uintptr_t segmentBase[16];
     
     // DMEM vertex buffer
     BKVertex dmem[DMEM_VERTEX_COUNT];
